DROID-4397 Extract ObservePersonalFavoriteTargets helper
PersonalFavoritesWidgetContainer no longer owns the block-tree
observation (OpenObject + InterceptEvents + reduce + target extraction).
All of that logic moves to a new domain-layer use case:

  ObservePersonalFavoriteTargets(space): Flow<List<Id>>

Opens the per-user _personalWidgets_<encodedSpaceId> doc, reduces the
relevant block events (AddBlock / DeleteBlock / UpdateStructure /
LinkGranularChange), walks root children, filters out built-in targets
(favorite / recent / bin / allObjects / chat / personalFavorites),
emits ordered real-object targets.

This unblocks Task 19d-part-2 — ObjectMenuOptionsProviderImpl will take
the same Flow (mapped to a Set) to drive the Favorite ↔ Unfavorite
menu item for any object screen, not just the home widgets view.

Container and delegate are shrunk: PersonalFavoritesWidgetContainer now
only handles the flat-list → WidgetView projection + object
subscription join; WidgetContainerDelegateImpl drops OpenObject +
InterceptEvents ctor args in favour of the single helper.
HomeScreenViewModel + its Factory pick up the helper via constructor.

Tests:
- 3 new ObservePersonalFavoriteTargetsTest cases (ordered emission,
  built-in filter, empty tree) — 3/3 pass.
- PersonalFavoritesWidgetContainerTest rewritten to mock the helper
  directly — 4/4 pass.
- HomeScreenViewModelTest updated to provide the new mock.

App compiles green, presentation + domain tests pass.
